Apply material

+++ frontend/src/app/pages/point-view/point-view.component.ts
@@ -1,31 +1,36 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { Component, Inject, OnInit, ViewChild } from '@angular/core';
+import { animate, state, style, transition, trigger } from '@angular/animations';
 import { MatTable } from '@angular/material/table';
 import { MatDialog } from '@angular/material/dialog';
-import { NgbModal, NgbModalOptions } from '@ng-bootstrap/ng-bootstrap';
 
 import Point from 'src/app/models/point';
 import { PointService } from 'src/app/point.service';
+import { AddPointDialog } from '../components/add-point-dialog';
+import { DeletePointDialog } from '../components/delete-point-dialog';
 
 @Component({
   selector: 'app-point-view',
   templateUrl: './point-view.component.html',
-  styleUrls: ['./point-view.component.css']
+  styleUrls: ['./point-view.component.css'],
+  animations: [
+    trigger('detailExpand', [
+      state('collapsed', style({height: '0px', minHeight: '0'})),
+      state('expanded', style({height: '*'})),
+      transition('expanded <=> collapsed', animate('225ms cubic-bezier(0.4, 0.0, 0.2, 1)')),
+    ]),
+  ]
 })
 export class PointViewComponent implements OnInit {
-  displayedColumns: string[] = ['code', 'name', 'partOfEar', 'bodyPart', 'function', 'action'];
+  columnsToDisplay = ['code', 'name', 'partOfEar', 'bodyPart', 'action'];
+  expandedPoint: Point | null;
   points: Point[] = [];
-  modalOptions: NgbModalOptions;
+
   @ViewChild(MatTable, { static: true }) table: MatTable<any>;
 
   constructor(
     public dialog: MatDialog,
-    private pointService: PointService,
-    private modalService: NgbModal
+    private pointService: PointService
   ) { 
-    this.modalOptions = {
-      backdrop:'static',
-      backdropClass:'customBackdrop'
-    }   
   }
 
   ngOnInit(): void {
@@ -33,30 +38,35 @@ export class PointViewComponent implements OnInit {
       .subscribe((points: Point[]) => this.points = points);
   }
 
-  addPoint(value: Point) {
-    this.pointService.createPoint(value)
-      .subscribe((points: Point[]) => {
-        this.pointService.getPoints();
-      });
-  }
+  openAddNewDialog() {
+    const dialogRef = this.dialog.open(AddPointDialog, {
+      width: '450px',
+      disableClose: true,
+      data: new Point
+    });
 
-  deletePoint(point: Point) {
-    this.pointService.deletePoint(point._id)
-      .subscribe(() => this.points = this.points.filter(l => l._id != point._id));
-  }
+    dialogRef.afterClosed().subscribe(result => {
+      console.log(`Dialog result: ${result.name}`);
 
-  open(content: any) {
-    this.modalService.open(content, this.modalOptions).result.then((result: Point) => {
       this.pointService.createPoint(result)
       .subscribe(() => this.pointService.getPoints()
       .subscribe((points: Point[]) => this.points = points));
-    }, (reason) => {
-      
     });
   }
 
-  openDialog(action, obj) {
-    obj.action = action;
+  openDeleteDialog(point: Point) {
+    const dialogRef = this.dialog.open(DeletePointDialog, {
+      width: '400px',
+      disableClose: true,
+      data: point
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result) {
+        this.pointService.deletePoint(point._id)
+        .subscribe(() => this.points = this.points.filter(l => l._id != point._id));
+      }
+    });
   }
 
   addRowData(row_obj: Point){
